+++ HISTORY.rst
@@ -43,6 +43,7 @@ End-User Summary
 - Added link-out to variant when present in ClinVar.
 - Fixed broken SV filter button in smallvariant filter form.
 - Added link-out to case from import bg job detail page.
+- Added ``recessive`` quick presets setting.
 
 Full Change List
 ================
@@ -99,6 +100,7 @@ Full Change List
 - Added link-out to variant when present in ClinVar by adding the SCV field from the HGNC database to the query.
 - Fixed broken SV filter button in smallvariant filter form.
 - Added link-out to case from import bg job detail page.
+- Added ``recessive`` quick presets setting.
 
 -------
 v0.18.0